+++ LibCarla/source/carla/sensor/s11n/NoopSerializer.cpp
@@ -4,17 +4,16 @@
 // This work is licensed under the terms of the MIT license.
 // For a copy, see <https://opensource.org/licenses/MIT>.
 
-#include "carla/sensor/s11n/LidarSerializer.h"
+#include "carla/sensor/s11n/NoopSerializer.h"
 
-#include "carla/sensor/data/LidarMeasurement.h"
+#include "carla/Exception.h"
 
 namespace carla {
 namespace sensor {
 namespace s11n {
 
-  SharedPtr<SensorData> LidarSerializer::Deserialize(RawData &&data) {
-    return SharedPtr<data::LidarMeasurement>(
-        new data::LidarMeasurement{std::move(data)});
+  SharedPtr<SensorData> NoopSerializer::Deserialize(RawData &&) {
+    throw_exception(std::runtime_error("NoopSerializer: Invalid data received."));
   }
 
 } // namespace s11n