+++ Unreal/CarlaUE4/Plugins/Carla/Source/Carla/Settings/CarlaSettings.h
@@ -85,10 +85,6 @@ public:
   /// Optional setting for the secondary port.
   TOptional<uint32> StreamingPort;
 
-  /// Time-out in milliseconds for the networking operations.
-  UPROPERTY(Category = "CARLA Server", VisibleAnywhere, meta = (EditCondition = bUseNetworking))
-  uint32 ServerTimeOut = 10000u;
-
   /// In synchronous mode, CARLA waits every tick until the control from the
   /// client is received.
   UPROPERTY(Category = "CARLA Server", VisibleAnywhere, meta = (EditCondition = bUseNetworking))