+++ Unreal/CarlaUE4/Plugins/Carla/Source/Carla/Vehicle/WheeledVehicleAIController.cpp
@@ -94,7 +94,7 @@ AWheeledVehicleAIController::AWheeledVehicleAIController(const FObjectInitialize
 AWheeledVehicleAIController::~AWheeledVehicleAIController() {}
 
 // =============================================================================
-// -- APlayerController --------------------------------------------------------
+// -- AController --------------------------------------------------------------
 // =============================================================================
 
 void AWheeledVehicleAIController::OnPossess(APawn *aPawn)
@@ -135,11 +135,9 @@ void AWheeledVehicleAIController::Tick(const float DeltaTime)
     return;
   }
 
-  TickAutopilotController();
-
   if (bAutopilotEnabled)
   {
-    Vehicle->ApplyVehicleControl(AutopilotControl, EVehicleInputPriority::Autopilot);
+    Vehicle->ApplyVehicleControl(TickAutopilotController(), EVehicleInputPriority::Autopilot);
   }
   else if (!bControlIsSticky)
   {
@@ -153,21 +151,25 @@ void AWheeledVehicleAIController::Tick(const float DeltaTime)
 // -- Autopilot ----------------------------------------------------------------
 // =============================================================================
 
-void AWheeledVehicleAIController::ConfigureAutopilot(const bool Enable)
+void AWheeledVehicleAIController::ConfigureAutopilot(const bool Enable, const bool KeepState)
 {
   bAutopilotEnabled = Enable;
-  // Reset state.
-  Vehicle->SetSteeringInput(0.0f);
-  Vehicle->SetThrottleInput(0.0f);
-  Vehicle->SetBrakeInput(0.0f);
-  Vehicle->SetReverse(false);
-  Vehicle->SetHandbrakeInput(false);
+  if (!KeepState)
+  {
+    // Reset state.
+    Vehicle->SetSteeringInput(0.0f);
+    Vehicle->SetThrottleInput(0.0f);
+    Vehicle->SetBrakeInput(0.0f);
+    Vehicle->SetReverse(false);
+    Vehicle->SetHandbrakeInput(false);
+    ClearQueue(TargetLocations);
+    Vehicle->SetAIVehicleState(
+        bAutopilotEnabled ?
+        ECarlaWheeledVehicleState::FreeDriving :
+        ECarlaWheeledVehicleState::AutopilotOff);
+  }
+
   TrafficLightState = ETrafficLightState::Green;
-  ClearQueue(TargetLocations);
-  Vehicle->SetAIVehicleState(
-      bAutopilotEnabled ?
-      ECarlaWheeledVehicleState::FreeDriving :
-      ECarlaWheeledVehicleState::AutopilotOff);
 
   /// @todo Workaround for a race condition between client and server when
   /// enabling autopilot right after initializing a vehicle.
@@ -215,13 +217,13 @@ void AWheeledVehicleAIController::SetFixedRoute(
 // -- AI -----------------------------------------------------------------------
 // =============================================================================
 
-void AWheeledVehicleAIController::TickAutopilotController()
+FVehicleControl AWheeledVehicleAIController::TickAutopilotController()
 {
 #if WITH_EDITOR // This happens in simulation mode in editor.
   if (Vehicle == nullptr)
   {
     bAutopilotEnabled = false;
-    return;
+    return {};
   }
 #endif // WITH_EDITOR
 
@@ -259,6 +261,8 @@ void AWheeledVehicleAIController::TickAutopilotController()
     Throttle = Move(Speed);
   }
 
+  FVehicleControl AutopilotControl;
+
   if (Throttle < 0.001f)
   {
     AutopilotControl.Brake = 1.0f;
@@ -270,6 +274,8 @@ void AWheeledVehicleAIController::TickAutopilotController()
     AutopilotControl.Throttle = Throttle;
   }
   AutopilotControl.Steer = Steering;
+
+  return AutopilotControl;
 }
 
 float AWheeledVehicleAIController::GoToNextTargetLocation(FVector &Direction)