+++ Unreal/CarlaUE4/Plugins/Carla/Source/Carla/Vehicle/WheeledVehicleAIController.h
@@ -8,7 +8,7 @@
 
 #include <queue>
 
-#include "GameFramework/PlayerController.h"
+#include "GameFramework/Controller.h"
 
 #include "Traffic/TrafficLightState.h"
 #include "Vehicle/VehicleControl.h"
@@ -21,7 +21,7 @@ class URoadMap;
 
 /// Wheeled vehicle controller with optional AI.
 UCLASS()
-class CARLA_API AWheeledVehicleAIController : public APlayerController
+class CARLA_API AWheeledVehicleAIController final : public AController
 {
   GENERATED_BODY()
 
@@ -38,7 +38,7 @@ public:
 
   /// @}
   // ===========================================================================
-  /// @name APlayerController overrides
+  /// @name Controller overrides
   // ===========================================================================
   /// @{
 
@@ -73,12 +73,6 @@ public:
   const ACarlaWheeledVehicle *GetPossessedVehicle() const
   {
     return Vehicle;
-  }
-
-  UFUNCTION(Category = "Wheeled Vehicle Controller", BlueprintCallable)
-  virtual bool IsPossessingThePlayer() const
-  {
-    return false;
   }
 
   /// @}
@@ -143,11 +137,11 @@ public:
   }
 
   UFUNCTION(Category = "Wheeled Vehicle Controller", BlueprintCallable)
-  void SetAutopilot(bool Enable)
+  void SetAutopilot(bool Enable, bool KeepState = false)
   {
     if (IsAutopilotEnabled() != Enable)
     {
-      ConfigureAutopilot(Enable);
+      ConfigureAutopilot(Enable, KeepState);
     }
   }
 
@@ -159,7 +153,7 @@ public:
 
 private:
 
-  void ConfigureAutopilot(bool Enable);
+  void ConfigureAutopilot(const bool Enable, const bool KeepState = false);
 
   /// @}
   // ===========================================================================
@@ -216,20 +210,10 @@ public:
   void SetFixedRoute(const TArray<FVector> &Locations, bool bOverwriteCurrent = true);
 
   /// @}
-  // ===========================================================================
-  /// @name AI
-  // ===========================================================================
-  /// @{
-
-  UFUNCTION(Category = "Wheeled Vehicle Controller", BlueprintCallable)
-  const FVehicleControl &GetAutopilotControl() const
-  {
-    return AutopilotControl;
-  }
 
 private:
 
-  void TickAutopilotController();
+  FVehicleControl TickAutopilotController();
 
   /// Returns steering value.
   float GoToNextTargetLocation(FVector &Direction);
@@ -277,8 +261,6 @@ private:
 
   UPROPERTY()
   ATrafficLightBase *TrafficLight;
-
-  FVehicleControl AutopilotControl;
 
   std::queue<FVector> TargetLocations;
 };